Rewrite rules reducer to use RTK
- Rewrote reducer to use `createSlice`
- Moved rules state typedefs into the reducer
- Deleted dead actions
