+++ src/devtools/client/inspector/rules/components/Rules.tsx
@@ -1,6 +1,6 @@
 import React from "react";
 import { Rule } from "devtools/client/inspector/rules/components/Rule";
-import { RuleState } from "../state/rules";
+import { RuleState } from "../reducers/rules";
 
 type RulesProps = {
   rules: RuleState[];
